requirements.txt
setuptools~=58.2.0
python-dotenv~=0.19.1
spotipy~=2.19.0
mutagen~=1.45.1
tqdm~=4.62.3
